Made a seperate file for the Pothole completed, so it is easier to refer in the future.
